legacy/backend/requirements.txt
fastapi==0.133.1
uvicorn==0.41.0
pydantic==2.12.5
pydantic-core==2.41.5
joblib==1.5.3
pandas==2.3.3
scikit-learn==1.8.0
requests==2.32.5
beautifulsoup4==4.14.3
tldextract==5.3.1
python-whois==0.9.6
dnspython==2.8.0
playwright==1.50.0
cachetools==6.2.6
numpy==2.4.2
scipy==1.17.1
threadpoolctl==3.6.0
python-dateutil==2.9.0.post0
six==1.17.0
pytz==2025.2
tzdata==2025.3
certifi==2026.2.25
charset-normalizer==3.4.4
idna==3.11
urllib3==2.6.3
soupsieve==2.8.3
h11==0.16.0
starlette==0.52.1
anyio==4.12.1
sniffio==1.3.1
typing-extensions==4.15.0
email-validator==2.2.0
